+++ libs/base/src/sys/mem/slice/mod.rs
@@ -3,11 +3,14 @@
 //! Slices of memory.
 //
 
-mod join; // join!
+mod join; // const_join!
 mod namespace; // Slice
 
 crate::structural_mods! { // _mods
     _mods {
-        pub use super::{namespace::Slice, join::join};
+        pub use super::{
+            namespace::*,
+            join::*,
+        };
     }
 }