+++ src/data/list/array/mod.rs
@@ -7,8 +7,6 @@
 //! They enable efficient iterable storage over a sequence of the same type.
 //
 
-crate::mod_path!(_c "../../../../libs/base/src/data/list/array/reexports.rs");
-
 mod adt; // DataArray
 mod d1; // 1-dimensional Array
 mod d2; // 2-dimensional Array2d
@@ -18,14 +16,20 @@ mod ext; // ExtArray, ArrayFmt
 #[cfg_attr(nightly_doc, doc(cfg(feature = "alloc")))]
 mod vec;
 
+// re-exports
+crate::mod_path!(_c "../../../../libs/base/src/data/list/array/reexports.rs");
+
 crate::structural_mods! { // _mods
     _mods {
-        pub use devela_base::data::list::{ArrayFrom, array_init};
-
-        pub use super::{_c::*, adt::*, d1::_all::*, d2::_all::*, ext::*};
+        pub use super::{adt::*, d1::_all::*, d2::_all::*, ext::*};
 
         #[cfg(feature = "alloc")]
         pub use super::vec::_all::*;
+
+        // re-exports
+        pub use super::_c::*;
+        #[doc(inline)]
+        pub use devela_base::data::list::{ArrayFrom, array_init};
     }
     _always {
         #[cfg(feature = "alloc")]