+++ src/num/float/reexports.rs
@@ -6,5 +6,5 @@
 crate::mod_path!(+pub _c "../../../libs/base/src/num/float/reexports.rs");
 
 // aliases
-#[doc = crate::TAG_DEVELA_BASE!()]
+#[doc(inline)]
 pub use devela_base::fsize;