+++ src/num/niche/mod.rs
@@ -53,6 +53,7 @@ mod impls; // impl ConstDefault, BitSized
 
 crate::structural_mods! { // _mods, _always
     _mods {
+        #[doc(inline)]
         pub use devela_base::num::niche::{
             NonValueU8, NonValueU16, NonValueU32, NonValueU64, NonValueU128, NonValueUsize,
             NonValueI8, NonValueI16, NonValueI32, NonValueI64, NonValueI128, NonValueIsize,