+++ src/num/ord/mod.rs
@@ -4,16 +4,15 @@
 // #![doc = crate::_doc!(extends: cmp)]
 // #![doc = crate::_doc!(modules: crate::num; cmp)]
 // #![doc = crate::_doc!(newline)]
-//
 
-// reexports
+// re-exports
 crate::mod_path!(_c "../../../libs/base/src/num/ord/reexports.rs");
 
 crate::structural_mods! { // _mods
     _mods {
+        // re-exports
+        pub use super::_c::*;
         #[doc(inline)]
         pub use devela_base::Compare;
-
-        pub use super::_c::*;
     }
 }