+++ src/phys/time/reexports.rs
@@ -6,6 +6,7 @@
 crate::mod_path!(+pub _c "../../../libs/base/src/phys/time/reexports.rs");
 crate::mod_path!(std +pub _s "../../../libs/base_std/src/phys/time/reexports.rs");
 
+#[doc(inline)]
 pub use devela_base::Timeout;
 
 #[cfg(feature = "std")]