+++ src/sys/mem/ptr/mod.rs
@@ -5,20 +5,20 @@
 // #![doc = crate::_doc!(extends: ptr)]
 //
 
-crate::mod_path!(_c "../../../../libs/base/src/sys/mem/ptr/reexports.rs");
-
 #[cfg(all(not(feature = "safe_mem"), feature = "unsafe_layout"))]
 mod fat;
 
+// re-exports
+crate::mod_path!(_c "../../../../libs/base/src/sys/mem/ptr/reexports.rs");
+
 crate::structural_mods! { // _mods, _always
     _mods {
-        pub use devela_base::sys::mem::Ptr;
-
-        pub use super::_c::*;
         #[cfg(all(not(feature = "safe_mem"), feature = "unsafe_layout"))]
         pub use super::fat::FatPtr;
-    }
-    _always {
+
+        // re-exports
         pub use super::_c::*;
+        #[doc(inline)]
+        pub use devela_base::sys::mem::Ptr;
     }
 }