+++ src/text/char.rs
@@ -11,8 +11,10 @@ crate::mod_path!(_c "../../libs/base/src/text/char/reexports.rs");
 
 crate::structural_mods! { // _mods
     _mods {
-        pub use devela_base::text::{Char, char7, char8, char16};
         pub use super::_c::*; // IterChars, char
+
+        #[doc(inline)]
+        pub use devela_base::text::{Char, char7, char8, char16};
     }
 }