+++ src/text/grapheme/mod.rs
@@ -12,10 +12,12 @@ mod r#trait; // Grapheme
 
 crate::structural_mods! { // _mods
     _mods {
+        pub use super::r#trait::*;
+
+        #[doc(inline)]
         pub use devela_base::text::{GraphemeNonul, GraphemeU8};
+        #[doc(inline)]
         #[cfg(feature = "alloc")]
         pub use devela_base_alloc::text::GraphemeString;
-
-        pub use super::r#trait::*;
     }
 }